Define AST structure(Node,Statement,Expression)
The Node interface is to be "implemented" by all types of nodes in
the AST.

Nodes in the AST which are statements(e.g. a  let statement) and nodes
in the AST which are expressions(5 * 5 * total) are to "implement" the
statement and expression interface respectively.

The Program struct represents the root node of the AST. It also stores
all the statements in the program.

The LetStatement struct represents a let statement and "implements" the
statement interface. It points to an identifier, a value which is to be
bound to that identifier and also stores its own token.

The identifier struct represents an identifier. It stores its own token
and also the value associated with this identifier.