Define the Object system
Object system defines how types in the interpreted
language are represented in the host programming language.

In Monkey, there are only three types, so far, and they are:
1. Integers
2. Booleans
3. Null

Every object in monkey is represented by a type in go that
confirms to the Object interface. Essentially, every type is
wrapped around in a struct that confirms to the Object interface.