Evaluate Return Statements
The evaluation of return statements requires the
separation of evaluation block statments into its
own function to correctly evaluate nested returns.

If a return statement is found in a non nested block,
that is currently, in the *ast.Program, its unwrapped
value is returned and execution stops.

If a return statement is found in a nested block, its
wrapped value is bubbled up all the way to the evaluator
of *ast.Program. This behaviour might change when we start
capturing return values into identifiers.